Initial landing zone
Created an initial landing zone with a Key Vault, Storage Account and Resource Group. This is a bootstrapped version. For this to work the "backend" tag under providers.tf must be commented out on an initial upload and once the SA and RG have been created these must be manually added in.